## Changes

- Optimize Keccak by ~3%

## Types of changes

#### What types of changes does your code introduce?

- [ ] Bugfix (a non-breaking change that fixes an issue)
- [ ] New feature (a non-breaking change that adds functionality)
- [ ] Breaking change (a change that causes existing functionality not to work as expected)
- [ ] Optimization
- [ ] Refactoring
- [ ] Documentation update
- [ ] Build-related changes
- [ ] Other: _Description_

## Testing

#### Requires testing

- [ ] Yes
- [x] No

#### If yes, did you write tests?

- [x] Yes
- [ ] No

## Remarks
Some benchmark runs:

| Method                | Mean     | Error   | StdDev  | Code Size | Allocated |
|---------------------- |---------:|--------:|--------:|----------:|----------:|
| From_span_keccak      | 241.2 ns | 4.50 ns | 4.82 ns |     758 B |         - |
| From_span_keccak_fast | 232.7 ns | 2.04 ns | 1.70 ns |     758 B |         - |
| From_span_keccak      | 241.5 ns | 3.31 ns | 2.59 ns |     758 B |         - |
| From_span_keccak_fast | 234.7 ns | 1.58 ns | 1.40 ns |     758 B |         - |
| From_span_keccak      | 236.8 ns | 0.40 ns | 0.37 ns |     758 B |         - |
| From_span_keccak_fast | 232.8 ns | 1.23 ns | 1.02 ns |     758 B |         - |
| From_span_keccak      | 238.0387 ns | 1.0068 ns | 0.9417 ns |     758 B |         - |
| From_span_keccak_fast | 230.8994 ns | 1.3858 ns | 1.2285 ns |     758 B |         - |
| From_span_keccak      | 237.1530 ns | 1.3545 ns | 1.1311 ns |     758 B |         - |
| From_span_keccak_fast | 229.3974 ns | 0.4813 ns | 0.4502 ns |     758 B |         - |
| From_span_keccak      | 238.5586 ns | 2.4952 ns | 2.2119 ns |     758 B |         - |
| From_span_keccak_fast | 233.5361 ns | 4.4838 ns | 4.6046 ns |     758 B |         - |