@Scooletz 2024-05-16
Can you compare asm and provide benches? Would be great to see the difference.